requirements.txt
python-telegram-bot==21.2
nest-asyncio==1.6.0